+++ src/main/kotlin/com/otakuexchange/domain/parimutuel/PortfolioPool.kt
@@ -13,9 +13,9 @@ data class PortfolioPool(
     val label: String,
     val entity: Entity? = null,
     val isWinner: Boolean,
-    val amount: Int,
+    val amount: Long,
     val volume: Long,
-    val userStake: Int?,  // null if user has no stake in this pool
+    val userStake: Long?,  // null if user has no stake in this pool
     val eventStatus: String,
     val createdAt: Instant = Clock.System.now(),
     val eventMultiplier: Int = 1
