+++ src/main/kotlin/com/otakuexchange/infra/tables/parimutuel/FirstStakeBonusTable.kt
@@ -7,7 +7,7 @@ import org.jetbrains.exposed.v1.core.Table
 object FirstStakeBonusTable : Table("first_stake_bonuses") {
     val userId  = uuid("user_id").references(UserTable.id)
     val eventId = uuid("event_id").references(EventTable.id)
-    val bonus   = integer("bonus")
+    val bonus   = long("bonus")
 
     override val primaryKey = PrimaryKey(userId, eventId)
 }
